requirements.txt
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.10.0
keras>=2.10.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
streamlit>=1.22.0
streamlit-canvas>=0.9.0
opencv-python>=4.6.0
pillow>=9.0.0
pandas>=1.4.0
plotly>=5.10.0
tqdm>=4.64.0
pygame>=2.1.0
